Changes to agitator for sserver and compactor
This change adds the `-a` argument to the service start command, just
like the accumulo-cluster script does for the Accumulo services. This
necessitated commenting out the datanode agitation as it might not
support the `-a` argument, or it might mean something else. Added
support for the sserver and compactor based on the cluster.yaml
configuration in this branch only.